requirements.txt
# Marketing Agency API Service Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Claude API
anthropic>=0.18.0

# Data validation
pydantic>=2.5.0

# HTTP client (for anthropic)
httpx>=0.26.0
gunicorn>=21.2.0

# Cloud & Utils
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
fal-client>=0.5.0
